move parseArgs short-group terminator tests to a .test.ts file
